+++ worker/src/RTC/Router.cpp
@@ -2,6 +2,9 @@
 // #define MS_LOG_DEV_LEVEL 3
 
 #include "RTC/Router.hpp"
+#ifdef MS_LIBURING_SUPPORTED
+#include "DepLibUring.hpp"
+#endif
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
 #include "Utils.hpp"
@@ -660,6 +663,11 @@ namespace RTC
 			// Clone only happens if needed.
 			std::shared_ptr<RTC::RtpPacket> sharedPacket;
 
+#ifdef MS_LIBURING_SUPPORTED
+			// Activate liburing usage.
+			DepLibUring::SetActive();
+#endif
+
 			for (auto* consumer : consumers)
 			{
 				// Update MID RTP extension value.
@@ -672,6 +680,11 @@ namespace RTC
 
 				consumer->SendRtpPacket(packet, sharedPacket);
 			}
+
+#ifdef MS_LIBURING_SUPPORTED
+			// Submit all prepared submission entries.
+			DepLibUring::Submit();
+#endif
 		}
 
 		auto it = this->mapProducerRtpObservers.find(producer);