+++ worker/src/RTC/RtpStreamSend.cpp
@@ -2,6 +2,9 @@
 // #define MS_LOG_DEV_LEVEL 3
 
 #include "RTC/RtpStreamSend.hpp"
+#ifdef MS_LIBURING_SUPPORTED
+#include "DepLibUring.hpp"
+#endif
 #include "Logger.hpp"
 #include "Utils.hpp"
 #include "RTC/RtpDictionaries.hpp"
@@ -125,6 +128,11 @@ namespace RTC
 
 		this->nackCount++;
 
+#ifdef MS_LIBURING_SUPPORTED
+		// Activate liburing usage.
+		DepLibUring::SetActive();
+#endif
+
 		for (auto it = nackPacket->Begin(); it != nackPacket->End(); ++it)
 		{
 			RTC::RTCP::FeedbackRtpNackItem* item = *it;
@@ -164,6 +172,11 @@ namespace RTC
 				}
 			}
 		}
+
+#ifdef MS_LIBURING_SUPPORTED
+		// Submit all prepared submission entries.
+		DepLibUring::Submit();
+#endif
 	}
 
 	void RtpStreamSend::ReceiveKeyFrameRequest(RTC::RTCP::FeedbackPs::MessageType messageType)
@@ -274,7 +287,7 @@ namespace RTC
 		return report;
 	}
 
-	RTC::RTCP::DelaySinceLastRr::SsrcInfo* RtpStreamSend::GetRtcpXrDelaySinceLastRr(uint64_t nowMs)
+	RTC::RTCP::DelaySinceLastRr::SsrcInfo* RtpStreamSend::GetRtcpXrDelaySinceLastRrSsrcInfo(uint64_t nowMs)
 	{
 		MS_TRACE();